feat: modularize cabinet configurator and scene
